HOCS moved from components/ to src/
