+++ src/components/offer/offer.jsx
@@ -7,9 +7,9 @@ import OfferImageList from '../offer-image-list/offer-image-list';
 import OfferFeatureList from '../offer-feature-list/offer-feature-list';
 import OfferHost from '../offer-host/offer-host';
 import OfferReviewList from '../offer-review-list/offer-review-list';
-import withOfferList from '../hocs/with-offer-list/with-offer-list';
+import withOfferList from '../../hocs/with-offer-list/with-offer-list';
 import OfferList from '../offer-list/offer-list';
-import withReviewForm from '../hocs/with-review-form/with-review-form';
+import withReviewForm from '../../hocs/with-review-form/with-review-form';
 import ReviewForm from '../review-form/review-form';
 import Map from '../map/map';
