+++ src/hocs/with-review-form/with-review-form.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 const withReviewForm = (Component) => {
