+++ src/hocs/with-sorting/with-sorting.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 const withSorting = (Component) => {
