+++ src/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {createStore} from "redux";
-import {Provider} from "react-redux";
+import {createStore} from 'redux';
+import {Provider} from 'react-redux';
 
 import App from './components/app/app';
 
 import mock from './mocks/offers';
-import {reducer} from "./store/reducer";
+import {reducer} from './store/reducer';
 
 const MockState = {
   rentalOffersCount: 32,
